Fixed a bug
When the user double clicks or clicks really fast on the same cards it messed up the game.

To prevent this I added a css class that will take away pointer events. I then added them to the card not matched section.
It checks how many cards are in the matchingCards array and disables mouse events accordingly.